+++ test/src/Fetcher/MissingItemIdFetcherTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace FactorioItemBrowserTest\Api\Search\Fetcher;
 
-use BluePsyduck\Common\Test\ReflectionTrait;
+use BluePsyduck\TestHelper\ReflectionTrait;
 use BluePsyduck\MapperManager\Exception\MapperException;
 use BluePsyduck\MapperManager\MapperManagerInterface;
 use FactorioItemBrowser\Api\Database\Entity\Item;
@@ -42,7 +42,6 @@ class MissingItemIdFetcherTest extends TestCase
 
     /**
      * Sets up the test case.
-     * @throws ReflectionException
      */
     protected function setUp(): void
     {
@@ -67,7 +66,6 @@ class MissingItemIdFetcherTest extends TestCase
 
     /**
      * Tests the fetch method.
-     * @throws ReflectionException
      * @throws MapperException
      * @covers ::fetch
      */