+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0 - 2019-07-09
+
+### Changed
+
+- Dependency `mtymek/blast-reflection-factory` with `bluepsyduck/zend-autowire-factory`.
+- Dependency `bluepsyduck/common` with `bluepsyduck/test-helper`.
+
 ## 1.0.0 - 2019-04-07
 
 - Initial release of the API search.